Separate LaunchProgram from mac implementation
